Fix clippy lints in migration code
- Remove unnecessary i32 cast in migrate_v4_to_v5
- Fix duplicate/malformatted doc comment on ensure_schema

+++ src/adapters/event_store/migrate_v4_to_v5.rs
@@ -64,7 +64,7 @@ impl Migration for MigrateV4ToV5 {
                 }
             } else {
                 // Blob entries (e.g., .schema-version) — keep as-is
-                root_builder.insert(entry_name, entry.id(), entry.filemode() as i32)?;
+                root_builder.insert(entry_name, entry.id(), entry.filemode())?;
             }
         }
